fix: write more unit tests and change ImportButton storybook to .tsx format
