+++ packages/components/stories/ImportButton.stories.tsx
@@ -1,16 +1,19 @@
 import React from 'react';
-
+import type { Meta, StoryObj } from '@storybook/react';
 import { ImportButton } from '../src/components/ImportButton';
 
-export default {
+const meta: any = {
   title: 'ImportButton',
   component: ImportButton,
   parameters: { layout: 'centered' },
   tags: ['autodocs'],
   argTypes: {},
-};
+} satisfies Meta<typeof ImportButton>;
 
-export const Default = {
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
   args: {
     label: 'Import',
     onChange: (data) => console.log(data)
